feat(token): make token literal optional for static values
This is mainly done as we do not get any new value from giving it on static tokens.
For example functions will always be set via the `fn` keyword, maybe later we have different
ways of doing it, but then the literal is probably not the correct place for more info.