feat(token): store position of the token for better logging
When working with the tokens etc. it will get more and more complicated to keep
an overview of the tokenized form. Instead of saying "hey we expected x but got y"
and needing to find out where this might be the case, now we can easily show the
input string and highlight the position the lexer was struggling at.

This is not perfect but better than not having any information at all