feat(lexer): handle new operators and refactor token creation
Mainly as otherwise the basic tokens have a gigantic list of arguments that are
just None, so now we do more of a builder pattern as.
Additionally the token itself should know its length and does already by knowing
which token it is, so keep this in the token implementation.