refactor(lexer): move literal initialization back to lexer
The main problem is, that if I have it interpreted in the token
then I have the same logic at multiple places.

Additionally this result in the token not knowing when it has what
and results to panic at compile time, I rather have more boilerplate
and then a better compile experience.